several changes
Shortened pointers, shortened logic, removed all but one of the declarations and replaced with hashset, changed deeppointer logic, removed entire update block so nothing runs every frame anymore.

I have not gotten any feedback, as it seems that nobody is willing to actually test this for me. It works for me and hope there are no errors, but it's impossible for me to test every scenario. Pointers for any version other than 12/18 are not 100% sure to work, but it's close to 90%.